Expose coupling transformer parameters
